feat: llc combine hook and context
