--- libs/ledger-live-common/src/platform/providers/LocalLiveAppProvider/types.ts
@@ -1,10 +1,0 @@
-import { LiveAppManifest } from "../../types";
-
-export type LiveAppRegistry = {
-  liveAppById: { [appId: string]: LiveAppManifest };
-  liveAppByIndex: LiveAppManifest[];
-};
-
-export type LiveAppDataAPI = {
-  fetchLiveAppManifests: (url: string) => Promise<LiveAppManifest[]>;
-};
