+++ libs/ledger-live-common/src/wallet-api/LocalLiveAppProvider/types.ts
@@ -1,0 +1,14 @@
+import { LiveAppManifest } from "../../platform/types";
+import { LocalLiveAppDB } from "../react";
+
+export type LiveAppProviderProps = {
+  children: React.ReactNode;
+  db: LocalLiveAppDB;
+};
+
+export type LiveAppContextType = {
+  state: LiveAppManifest[];
+  addLocalManifest: (LiveAppManifest) => void;
+  removeLocalManifestById: (string) => void;
+  getLocalLiveAppManifestById: (string) => LiveAppManifest | undefined;
+};
